+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 - `k8sclusterreceiver`: Validate that k8s API supports a resource before setting up a watcher for it (#9523)
 - `internal/stanza`: Add support for `remove` operator (#9524)
+- `googlecloudpubsubreceiver` Activate the Google Cloud Pubsub receiver. (#8853)
+- `googlecloudpubsubexporter` Activate the Google Cloud Pubsub exporter. (#8853)
 
 ### 🧰 Bug fixes 🧰